Migrate country api fetching and handling to modal file
