+++ client/src/routes/wall/+page.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
     import { v4 as uuid } from "uuid";
     import CountryCard from "$lib/components/CountryCard.svelte";
-    import { onMount } from "svelte";
-    import { fetchCountries } from "$lib/util/CountrySelectionService.svelte";
     import {
         Modal,
         initializeStores,
@@ -13,21 +11,6 @@
 
     initializeStores();
     const modalStore = getModalStore();
-
-    let selectedCountry = "";
-    let countries: Array<string> = [];
-
-    onMount(async () => {
-        try {
-            countries = await fetchCountries();
-        } catch (error) {
-            console.error("Error fetching countries:", error);
-        }
-    });
-
-    const handleCountryChange = (event: any) => {
-        selectedCountry = event.target.value;
-    };
 
     type CountryCardType = {
         id: string;
@@ -42,11 +25,6 @@
             title: "Showcase the country you visited!",
             component: {
                 ref: CountryPickerModal,
-                props: {
-                    countries,
-                    selectedCountry,
-                    handleCountryChange,
-                },
             },
             response: (selectedCountry: string) => {
                 if (selectedCountry) {
